+++ core-ng/src/main/java/core/framework/internal/web/sys/PropertyController.java
@@ -1,7 +1,7 @@
 package core.framework.internal.web.sys;
 
 import core.framework.internal.module.PropertyManager;
-import core.framework.internal.web.http.IPv4AccessControl;
+import core.framework.internal.web.http.IPAccessControl;
 import core.framework.web.Controller;
 import core.framework.web.Request;
 import core.framework.web.Response;
@@ -15,7 +15,7 @@ import java.util.TreeSet;
  * @author neo
  */
 public class PropertyController implements Controller {
-    private final IPv4AccessControl accessControl = new IPv4AccessControl();
+    private final IPAccessControl accessControl = new IPAccessControl();
     private final PropertyManager propertyManager;
 
     public PropertyController(PropertyManager propertyManager) {