+++ core-ng/src/test/java/core/framework/internal/web/http/IPAccessControlTest.java
@@ -8,18 +8,19 @@ import java.net.InetAddress;
 import java.net.UnknownHostException;
 import java.util.List;
 
+import static core.framework.internal.web.http.IPRanges.address;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 /**
  * @author neo
  */
-class IPv4AccessControlTest {
-    private IPv4AccessControl accessControl;
+class IPAccessControlTest {
+    private IPAccessControl accessControl;
 
     @BeforeEach
-    void createIPv4AccessControl() {
-        accessControl = new IPv4AccessControl();
+    void createIPAccessControl() {
+        accessControl = new IPAccessControl();
     }
 
     @Test
@@ -40,8 +41,8 @@ class IPv4AccessControlTest {
     void validateWithNotAllowedIP() {
         accessControl.allow = new IPv4Ranges(List.of("100.100.100.100/32"));
         assertThatThrownBy(() -> accessControl.validate("100.100.100.1"))
-                .isInstanceOf(ForbiddenException.class)
-                .hasMessageContaining("access denied");
+            .isInstanceOf(ForbiddenException.class)
+            .hasMessageContaining("access denied");
     }
 
     @Test
@@ -56,42 +57,36 @@ class IPv4AccessControlTest {
 
     @Test
     void allowWithBothAllowDenyDefined() {
-        accessControl.deny = new IPv4Ranges(List.of("100.100.100.100/24"));
-        accessControl.allow = new IPv4Ranges(List.of("100.100.100.101/32"));
+        var deny = new IPv4Ranges(List.of("100.100.100.100/24"));
+        var allow = new IPv4Ranges(List.of("100.100.100.101/32"));
 
-        assertThat(accessControl.allow(IPv4Ranges.address("100.100.99.1"))).isTrue();
-        assertThat(accessControl.allow(IPv4Ranges.address("100.100.100.101"))).isTrue();
-        assertThat(accessControl.allow(IPv4Ranges.address("100.100.100.1"))).isFalse();
+        assertThat(accessControl.allow(address("100.100.99.1"), allow, deny)).isTrue();
+        assertThat(accessControl.allow(address("100.100.100.101"), allow, deny)).isTrue();
+        assertThat(accessControl.allow(address("100.100.100.1"), allow, deny)).isFalse();
     }
 
     @Test
     void allowWithOnlyAllowDefined() {
-        accessControl.allow = new IPv4Ranges(List.of("100.100.100.101/32"));
+        var allow = new IPv4Ranges(List.of("100.100.100.101/32"));
 
-        assertThat(accessControl.allow(IPv4Ranges.address("100.100.99.1"))).isFalse();
-        assertThat(accessControl.allow(IPv4Ranges.address("100.100.100.101"))).isTrue();
-        assertThat(accessControl.allow(IPv4Ranges.address("100.100.100.1"))).isFalse();
+        assertThat(accessControl.allow(address("100.100.99.1"), allow, null)).isFalse();
+        assertThat(accessControl.allow(address("100.100.100.101"), allow, null)).isTrue();
+        assertThat(accessControl.allow(address("100.100.100.1"), allow, null)).isFalse();
     }
 
     @Test
     void allowWithOnlyDenyDefined() {
-        accessControl.deny = new IPv4Ranges(List.of("100.100.100.101/32"));
+        var deny = new IPv4Ranges(List.of("100.100.100.101/32"));
 
-        assertThat(accessControl.allow(IPv4Ranges.address("100.100.99.1"))).isTrue();
-        assertThat(accessControl.allow(IPv4Ranges.address("100.100.100.101"))).isFalse();
-        assertThat(accessControl.allow(IPv4Ranges.address("100.100.100.1"))).isTrue();
+        assertThat(accessControl.allow(address("100.100.99.1"), null, deny)).isTrue();
+        assertThat(accessControl.allow(address("100.100.100.101"), null, deny)).isFalse();
+        assertThat(accessControl.allow(address("100.100.100.1"), null, deny)).isTrue();
     }
 
     @Test
     void denyByDefault() {
-        accessControl.allow = null;
-        accessControl.deny = null;
+        assertThat(accessControl.allow(address("100.100.100.100"), null, null)).isFalse();
 
-        assertThat(accessControl.allow(IPv4Ranges.address("100.100.100.100"))).isFalse();
-    }
-
-    @Test
-    void allowWithIPv6() {
-        assertThat(accessControl.allow(IPv4Ranges.address("2001:0db8:85a3:0000:0000:8a2e:0370:0000"))).isTrue();
+        assertThat(accessControl.allow(address("2001:0db8:85a3:0000:0000:8a2e:0370:0000"), null, null)).isFalse();
     }
 }