+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### 9.2.2 (5/21/2025 - )
 
 * http_client: support proxy
+* ip: support ipv6
+  > support ipv6 ranges in access control
+  > !!! IPRangePropertyValueParser parsing behavior slightly changed, "NAME: CIDR1, CIDR2;", the "NAME:" must have space after it
 
 ### 9.2.1 (4/24/2025 - 5/19/2025) !!! only support java 24